Refactor endpoint validation state helpers
